Integrated Exceptions and updated services and repositories

+++ src/main/java/org/backendprojectsst/hotelmanagementsystem/controllers/RoomController.java
@@ -1,5 +1,7 @@
 package org.backendprojectsst.hotelmanagementsystem.controllers;
 
+import org.backendprojectsst.hotelmanagementsystem.exceptions.InvalidRoomIDException;
+import org.backendprojectsst.hotelmanagementsystem.exceptions.RoomAlreadyOccupiedException;
 import org.backendprojectsst.hotelmanagementsystem.models.Room;
 import org.backendprojectsst.hotelmanagementsystem.models.RoomStatus;
 import org.backendprojectsst.hotelmanagementsystem.models.RoomType;
@@ -27,8 +29,8 @@ public class RoomController
         return roomService.getAllRooms();
     }
 
-    @GetMapping("/status/{status}")
-    public List<Room> getRoomsByStatus(@PathVariable RoomStatus status)
+    @GetMapping("/status/")
+    public List<Room> getRoomsByStatus(@RequestParam RoomStatus status)
     {
         return roomService.getRoomsbyStatus(status);
     }
@@ -40,14 +42,16 @@ public class RoomController
     }
 
     @GetMapping("/{id}")
-    public Room getRoomById(@PathVariable long id)
+    public Room getRoomById(@PathVariable long id) throws InvalidRoomIDException
     {
+        if(!roomService.isRoom(id)) throw new InvalidRoomIDException("Invalid Room ID");
         return roomService.getRoomByCustomerId(id);
     }
 
     @GetMapping("/status/{id}")
-    public RoomStatus getRoomStatus(@PathVariable long id)
+    public RoomStatus getRoomStatus(@PathVariable long id) throws InvalidRoomIDException
     {
+        if(!roomService.isRoom(id)) throw new InvalidRoomIDException("Invalid Room ID");
         return roomService.getRoomStatus(id);
     }
 
@@ -58,32 +62,38 @@ public class RoomController
     }
 
     @PutMapping("/{id}/occupy/{customerId}")
-    public Room updateRoomOccupancy(@PathVariable long id, @PathVariable long customerId)
+    public Room updateRoomOccupancy(@PathVariable long id, @PathVariable long customerId) throws InvalidRoomIDException, RoomAlreadyOccupiedException
     {
+        if(!roomService.isRoom(id)) throw new InvalidRoomIDException("Invalid Room ID");
+        if(roomService.getRoomByCustomerId(id).getCustomer() != null) throw new RoomAlreadyOccupiedException("Room is already occupied");
         return roomService.updateRoomOccupancy(roomService.getRoomByCustomerId(id), customerId);
     }
 
     @PutMapping("/{id}/vacate")
-    public Room vacateRoom(@PathVariable long id)
+    public Room vacateRoom(@PathVariable long id) throws InvalidRoomIDException
     {
+        if(!roomService.isRoom(id)) throw new InvalidRoomIDException("Invalid Room ID");
         return roomService.updateRoomOccupancy(roomService.getRoomByCustomerId(id), null);
     }
 
     @PutMapping("/{id}/price")
-    public Room updateRoomPrice(@PathVariable long id, @RequestBody double price)
+    public Room updateRoomPrice(@PathVariable long id, @RequestBody double price) throws InvalidRoomIDException
     {
+        if(!roomService.isRoom(id)) throw new InvalidRoomIDException("Invalid Room ID");
         return roomService.updateRoomPrice(roomService.getRoomByCustomerId(id), price);
     }
 
     @PutMapping("/{id}/type")
-    public Room updateRoomType(@PathVariable long id, @RequestBody String type)
+    public Room updateRoomType(@PathVariable long id, @RequestBody String type) throws InvalidRoomIDException
     {
+        if(!roomService.isRoom(id)) throw new InvalidRoomIDException("Invalid Room ID");
         return roomService.updateRoomType(roomService.getRoomByCustomerId(id), type);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteRoom(@PathVariable long id)
+    public void deleteRoom(@PathVariable long id) throws InvalidRoomIDException
     {
+        if(!roomService.isRoom(id)) throw new InvalidRoomIDException("Invalid Room ID");
         roomService.deleteRoom(id);
     }
